requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
google-generativeai==0.2.1
